Full vulnScanningSync fixes
