+++ backend/src/xfd_django/xfd_api/api_methods/sync.py
@@ -14,7 +14,7 @@ from uuid import uuid4
 # Third-Party Libraries
 from django.db import transaction
 from fastapi import HTTPException, Request
-from xfd_api.tasks.vulnScanningSync import save_organization_to_mdl
+from xfd_api.tasks.utils.vs_requests import save_organization_to_mdl
 from xfd_mini_dl.models import Organization, Sector
 
 from ..auth import is_global_view_admin
