+++ backend/src/xfd_django/xfd_api/schema_models/scan.py
@@ -204,8 +204,8 @@ SCAN_SCHEMA = {
         type="fargate",
         is_passive=True,
         global_scan=True,
-        cpu="4096",
-        memory="16384",
+        cpu="16384",
+        memory="65536",
         description="Pull in vulnerability data from VSs Vulnerability database",
     ),
     "cveSync": ScanSchema(
